Add tests for get_overview_value_representation
